+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### BUG FIXES
 
-- `[evidence]` test(evidence): fix flaky `TestReactorsGossipNoCommittedEvidence`
+- `[evidence]` fix flaky `TestReactorsGossipNoCommittedEvidence` test
   ([\#5870](https://github.com/cometbft/cometbft/pull/5870))
 - `[blocksync]` fix removeTimedoutPeers deadlock found via Byzantine prevote gossip race
   ([\#5839](https://github.com/cometbft/cometbft/pull/5839))